Implement RFC 3139: alternative registry authentication support
Allows registries to request Cargo to send the authentication token
for all requests, rather than just publish/yank.

In the case of HTTP registries, Cargo first attempts an un-authenticated
request, then if that fails with HTTP 401, an authenticated request is attempted.
The registry server may include a `www-authenticate` header with the HTTP 401
to instruct Cargo with URL the user can visit to acquire a token (crates.io/me).

Since the API URL is not known (because it's stored in the index), the
unstable credential provider feature is modified to key off the index url,
and the registry name is no longer provided.

To handle the case where an alternative registry's name is not known
(such as coming from a lock file, or via --index), Cargo can now look up
the token in the configuration by matching on the index URL. This introduces
a new error if two alternative registries are configured with the same
index URL.

Several operations, such as `cargo install` could have had a `--token`
argument added, however it appears that Cargo would like to move away from
passing the token on the command line for security reasons. In this case,
users would need to configure the registry via the config file (or
environment variables) when using `cargo install --index ...` or similar.